--- pagingflowlibrary/src/commonMain/kotlin/ru/snowmaze/pagingflow/presenters/composite/CompositePagingDataPresenter.kt
@@ -1,6 +1,0 @@
-package ru.snowmaze.pagingflow.presenters.composite
-
-// TODO это должно быть именно презентером, т.к. сложно будет сделать это же через медиум да иэто не нужно
-internal class CompositePagingDataPresenter {
-
-}
